requirements.txt
cffi==1.17.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.1
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.2.6
packaging==25.0
pillow==11.3.0
pycparser==2.22
pygame-ce==2.5.5
pymunk==6.10.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyTMX==3.32
scipy==1.15.3
six==1.17.0
